refactor: get observation class from path
